% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdseyeview-package.R
\docType{package}
\name{birdseyeview-package}
\alias{birdseyeview}
\alias{birdseyeview-package}
\title{birdseyeview: Make Planning Maps, Tables, and Visualizations}
\description{
This package helps create maps, tables, and figures for planning documents using open spatial data. birdseyeview is designed to work with the overedge package and relies on sf, ggplot2, and gt.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elipousson.github.io/birdseyeview/}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID})

}
\keyword{internal}
