% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markers.R
\name{get_markers}
\alias{get_markers}
\title{Get markers}
\usage{
get_markers(
  data,
  groupname_col = "group",
  group_meta = NULL,
  group_join = sf::st_intersects,
  geocode = FALSE,
  address_col = "address",
  point = TRUE,
  crs = NULL,
  fn = NULL,
  ...
)
}
\arguments{
\item{data}{Data with markers, passed to data parameter of `overedge::get_location_data`}

\item{groupname_col}{Group column name, used to join group metadata if group_meta is a non-spatial data frame; Default: 'group'}

\item{group_meta}{Group metadata as a data frame or sf object that intersect with markers; Default: NULL}

\item{group_join}{The join function used by sf::st_join group_meta is an sf object, Default: sf::st_intersects()}

\item{geocode}{If TRUE, geocode data using `tidygeocoder::geo` and then convert to sf with `overedge::df_to_sf`, Default: FALSE}

\item{address_col}{Address column, used if geocode is TRUE Default: 'address'}

\item{point}{If TRUE, convert geometry to POINT with sf::st_centroid(), Default: TRUE}

\item{crs}{Coordinate reference system for markers, Default: NULL}

\item{fn}{Function to apply to data before results; gives warning if data is grouped; Default: NULL}

\item{...}{Additional parameters passed to `overedge::get_location_data`}
}
\description{
Get markers
}
