% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_geom_sf_rows.R
\name{gt_geom_sf_rows}
\alias{gt_geom_sf_rows}
\title{Add maps of simple features into rows of a `gt` table}
\usage{
gt_geom_sf_rows(
  gt_object,
  mapping = NULL,
  fill = NA,
  color = "black",
  size = 1,
  linetype = "solid",
  height = 100,
  asp = 1,
  fn = NULL,
  ...
)
}
\arguments{
\item{gt_object}{An existing gt table object of class gt_tbl. The gt_object must be convertible to an sf object with sf::st_as_sf so do not drop the geometry before creating an sf object.}

\item{mapping}{aesthetic mapping passed to layer_location_data, Default: NULL}

\item{fill}{Fill (used for polygon geometry), Default: NA}

\item{color}{Color (used for polygon, point, and line geometry), Default: 'black'}

\item{size}{Color (used for polygon, point, and line geometry), Default: 1}

\item{linetype}{Color (used for polygon and line geometry), Default: 'solid'}

\item{height}{The absolute height (px) of the map in the table cell}

\item{asp}{The aspect ratio of the map in the table cell.}

\item{fn}{Optional rowwise function that results in ggplot2 map the replaces geometry column.}
}
\description{
Add maps of simple features into rows of a `gt` table with {ggplot2}. Function naming convention borrowed from {gtExtras} <https://jthomasmock.github.io/gtExtras/reference/gt_img_rows.html>
}
\concept{gt}
