% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_merge_stack.R
\name{gt_merge_stack}
\alias{gt_merge_stack}
\title{Merge and stack text from two columns in \code{gt} (variation on \code{{gtExtras}})}
\usage{
gt_merge_stack(
  gt_object,
  columns = c(1, 2),
  font_color = c("black", "grey"),
  ...,
  small_cap = FALSE,
  font_size = c(14, 10),
  font_weight = "bold",
  h_rule = FALSE
)
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}

\item{columns}{Columns to stack, Default: c(1, 2)}

\item{font_color}{Color(s) to use for stacked text. Default: c("black",
"grey")}

\item{...}{
  Arguments passed on to \code{\link[scales:col2hcl]{scales::col2hcl}}
  \describe{
    \item{\code{h}}{Hue, \verb{[0, 360]}}
    \item{\code{c}}{Chroma, \verb{[0, 100]}}
    \item{\code{l}}{Luminance, \verb{[0, 100]}}
    \item{\code{alpha}}{Alpha, \verb{[0, 1]}.}
  }}

\item{small_cap}{a logical indicating whether to use 'small-cap' on the top line of text}

\item{font_size}{a string of length 2 indicating the font-size in px of the top and bottom text}

\item{font_weight}{a string of length 2 indicating the 'font-weight' of the top and bottom text. Must be one of 'bold', 'normal', 'lighter'}

\item{h_rule}{Horizontal rule to separate top and bottom column text. If
\code{TRUE}, use a 1px line with a 2px margin on top and bottom. If \code{FALSE},
include no horizontal rule. To create a custom rule, provide a \verb{<hr>}
thematic break tag or a \verb{<div>} with css styling.}
}
\description{
A variation on the \link[gtExtras:gt_merge_stack]{gtExtras::gt_merge_stack} function that implements a
horizontal rule parameter separating the top and bottom text and simplifies
the requirements for the input variables. \code{{gtExtras}} was created by Thomas
Mock and is available under an MIT License.
}
