% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_sf_rows.R
\name{gt_sf_rows}
\alias{gt_sf_rows}
\alias{gt_geom_sf_rows}
\title{Add maps of simple features into rows of a \code{gt} table}
\usage{
gt_sf_rows(
  gt_object,
  mapping = NULL,
  fill = NA,
  color = "black",
  size = 1,
  linetype = "solid",
  height = 100,
  asp = 1,
  geom = "sf",
  layer_fn = NULL,
  sf_label = "map",
  ...
)
}
\arguments{
\item{gt_object}{An existing gt table object of class gt_tbl or an sf object.
If a gt_object must be convertible to an sf object with \code{\link[sf:st_as_sf]{sf::st_as_sf()}}
(e.g. geometry must be retained before passing to \code{\link[gt:gt]{gt::gt()}}).}

\item{mapping}{aesthetic mapping passed to \code{\link[maplayer:layer_location_data]{maplayer::layer_location_data()}}, Default: \code{NULL}}

\item{fill}{Fill (used for polygon geometry), Default: \code{NA}}

\item{color}{Color (used for polygon, point, and line geometry), Default:
'black'}

\item{size}{Color (used for polygon, point, and line geometry), Default: 1}

\item{linetype}{Color (used for polygon and line geometry), Default: 'solid'}

\item{height}{The absolute height (px) of the map in the table cell}

\item{asp}{The aspect ratio of the map in the table cell.}

\item{layer_fn}{Optional rowwise function that results in ggplot2 map the replaces
geometry column.}

\item{sf_label}{Label used for column with ggplot2 map.}
}
\description{
Add maps of simple features into rows of a \code{gt} table with {ggplot2}.
Function naming convention borrowed from \link[gtExtras:gt_img_rows]{gtExtras::gt_img_rows}. Previously
named gt_geom_sf_rows.
}
\concept{gt}
