% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inset_map.R
\name{make_inset_map}
\alias{make_inset_map}
\title{Create a map with an inset context map using ggplot2}
\usage{
make_inset_map(
  map = NULL,
  inset = NULL,
  location = NULL,
  context = NULL,
  position = "bottomright",
  nudge_x = 0,
  nudge_y = 0
)
}
\arguments{
\item{map}{plot or map created with ggplot2}

\item{inset}{plot or map created with ggplot2}

\item{location}{sf or bbox object}

\item{context}{sf or bbox object with location context, Default: NULL}

\item{position}{inset map position, Default: 'bottomright'}

\item{nudge_x}{nudge X position of inset map, Default: 0}

\item{nudge_y}{nudge Y position of inset map, Default: 0}
}
\value{
ggplot2 with inset map added using patchwork
}
\description{
Works with `layer_location_context`
}
\examples{
\dontrun{
if (interactive()) {
  map <-
    ggplot2::ggplot() +
    overedge::layer_location_data(data = mapbaltimore::get_area("neighborhood", "Harwood")) +
    overedge::layer_neatline(
      data = mapbaltimore::get_area("neighborhood", "Harwood"),
      asp = "8.5:11"
    )

  location <-
    mapbaltimore::get_area("neighborhood", "Harwood")

  context <-
    mapbaltimore::baltimore_city

  inset_context_map(
    map = map,
    location = location,
    context = context,
    position = "bottomright",
    nudge_x = -0.05,
    nudge_y = 0.05
  )
}
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}
 \code{\link[patchwork]{inset_element}}
}
