% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_group_data.R
\name{layer_group_data}
\alias{layer_group_data}
\title{Make group layers}
\usage{
layer_group_data(
  data,
  mapping = NULL,
  groupname_col = "group",
  geom = "sf",
  layers = TRUE,
  ...
)
}
\arguments{
\item{data}{sf object including data in area. data may also be a url or file
path. data can be the name of a data object or, if package and filetype are
provided, a cached or external file.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{groupname_col}{Group column name}

\item{geom}{ggplot2 geom to use, Default: 'sf'. Options include "sf"
(geom_sf), "icon" (geom_sf_icon), "text" (geom_sf_text), "label"
(geom_sf_label), "textsf", "labelsf", "text_repel", and "label_repel".}

\item{layers}{If `TRUE`; return a list of layers; if `FALSE`; return a list
of ggplot2 maps; defaults to `TRUE`}

\item{...}{Additional parameters passed to `overedge::layer_location_data`}
}
\description{
Make group layers
}
