% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_location_context.R
\name{layer_location_context}
\alias{layer_location_context}
\title{Create a map layer showing a location in context of another area or location}
\usage{
layer_location_context(
  data = NULL,
  fill = "gray70",
  color = "black",
  context = NULL,
  context_aes = list(fill = "white", color = "black", alpha = 1, ...),
  crs = NULL,
  neatline = TRUE,
  ...
)
}
\arguments{
\item{data}{sf object including data in area. data may also be a url or file
path. data can be the name of a data object or, if package and filetype are
provided, a cached or external file.}

\item{context}{sf or bbox object for context area or a `geom` layer}

\item{context_aes}{list with aesthetic attributes for context area; must include fill and color; defaults to list(fill = "white", color = "black", ...)}

\item{crs}{coordinate reference system to return}

\item{neatline}{If `TRUE`, add a neatline layer to the returned layer}

\item{...}{Parameters passed to selected geom}
}
\description{
Intended for use with inset_map function.
}
