% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_show_markers.R
\name{layer_show_markers}
\alias{layer_show_markers}
\alias{layer_number_markers}
\title{Add a marker layer to a map with or without numbered markers}
\usage{
layer_show_markers(
  data,
  mapping = NULL,
  geom = "sf",
  style = NULL,
  get = FALSE,
  groupname_col = NULL,
  group_meta = NULL,
  crs = NULL,
  fn = NULL,
  number = FALSE,
  number_col = NULL,
  scale = NULL,
  ...
)

layer_number_markers(
  data,
  mapping = NULL,
  number_col = NULL,
  groupname_col = NULL,
  number_by = FALSE,
  size = 5,
  style = "roundrect",
  geom = "label",
  sort = "lon",
  desc = NULL,
  scale = NULL,
  ...
)
}
\arguments{
\item{data}{Data with markers, passed to data parameter of
\code{\link[overedge:get_location_data]{get_location_data()}}}

\item{style}{Style; defaults to \code{NULL} for \code{\link[=layer_show_markers]{layer_show_markers()}} (supports
"facet"); defaults to "roundrect" for \code{\link[=layer_number_markers]{layer_number_markers()}},}

\item{get}{If \code{TRUE}, pass data to \link[overedge:make_markers]{overedge::make_markers}}

\item{groupname_col}{Group column name, used to join group metadata if
group_meta is a non-spatial data frame; Default: \code{NULL}}

\item{group_meta}{Group metadata as a data frame or sf object that intersect
with markers; Default: \code{NULL}}

\item{crs}{Coordinate reference system for markers, Default: \code{NULL}}

\item{fn}{Function to apply to data before results; gives warning if data is
grouped; Default: \code{NULL}}

\item{number}{If \code{TRUE}, number markers using \code{\link[=layer_number_markers]{layer_number_markers()}} (not
currently supported)}

\item{number_col}{Name of column with numbers; defaults to NULL.}

\item{...}{Additional parameters passed to \code{overedge::layer_location_data}
(include label.size, label.padding, and label.r to define alternate style)}

\item{size}{Marker size, Default: 5}

\item{sort}{Sort column name, Default: "dist_xmin_ymax".}

\item{desc}{If \code{TRUE}, sort descending; default \code{FALSE}.}
}
\value{
ggplot2 layers
}
\description{
If get is \code{TRUE}, groupname_col, group_meta, crs, and fn is all passed on to
\link[overedge:make_markers]{overedge::make_markers}
}
\details{
The number parameter is not currently supported so the number_col parameter
is not implemented.
}
\examples{
\dontrun{
if (interactive()) {
  ggplot2::ggplot() +
    layer_number_markers(
      data = mapbaltimore::parks,
      groupname_col = "park_district"
    )
}
}
}
