% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_show_markers.R
\name{make_group_layers}
\alias{make_group_layers}
\title{Make group layers}
\usage{
make_group_layers(
  data,
  mapping = ggplot2::aes(),
  groupname_col = "group",
  layers = FALSE,
  ...
)
}
\arguments{
\item{data}{sf object including data in area. data may also be a url or file
path. data can be the name of a data object or, if package and filetype are
provided, a cached or external file.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{layers}{defaults to FALSE}

\item{...}{Parameters passed to selected geom}
}
\description{
Make group layers
}
