% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mapbox_basemap.R
\name{make_mapbox_basemap}
\alias{make_mapbox_basemap}
\alias{layer_show_mapbox}
\title{Use snapbox to make a Mapbox basemap}
\usage{
make_mapbox_basemap(
  data,
  map_style = snapbox::mapbox_gallery_moonlight(),
  scale_ratio = 0.5,
  access_token = Sys.getenv("MAPBOX_PUBLIC_TOKEN"),
  basemap = TRUE,
  neatline = TRUE,
  expand = TRUE,
  mapbox_logo = TRUE,
  attribution = TRUE,
  ...
)

layer_show_mapbox(
  data,
  map_style = snapbox::mapbox_gallery_moonlight(),
  scale_ratio = 0.5,
  access_token = Sys.getenv("MAPBOX_PUBLIC_TOKEN"),
  basemap = FALSE,
  neatline = FALSE,
  expand = FALSE,
  mapbox_logo = TRUE,
  attribution = TRUE,
  ...
)
}
\arguments{
\item{data}{sf, sfc, or bbox object; any objects convertible with \code{\link[overedge:as_sf]{overedge::as_bbox()}}}

\item{map_style}{Map style, Default: \code{\link[snapbox:reexports]{snapbox::mapbox_gallery_moonlight()}}}

\item{scale_ratio}{Scale ratio, Default: 1}

\item{access_token}{Access token, following default format for public token from \code{\link[mapboxapi:mb_access_token]{mapboxapi::mb_access_token()}}, Default: Sys.getenv("MAPBOX_PUBLIC_TOKEN")}

\item{basemap}{If FALSE, create a stand alone layer; if TRUE, the layer is precededed by \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} to allow use as a basempa, Default: TRUE}

\item{neatline}{If TRUE, add a neatline matching the provided data, Default: TRUE}

\item{expand}{Default \code{FALSE.} If \code{TRUE}, the function adds
\code{\link[ggplot2]{scale_y_continuous}} and
\code{\link[ggplot2]{scale_x_continuous}} to expand the map extent to
provided parameters.}

\item{...}{Additional parameter passed to \code{\link[overedge:layer_neatline]{overedge::layer_neatline()}}}
}
\description{
Expect this function to shift from using snapbox to mapboxapi if/when this
pull request is accepted: \url{https://github.com/walkerke/mapboxapi/pull/18}
}
\details{
layer_show_mapbox works the same as make_mapbox_basemap but defaults to
basemap and neatline to FALSE
}
\seealso{
\code{\link[snapbox]{reexports}},\code{\link[snapbox]{layer_mapbox}}
\code{\link[overedge]{st_transform_ext}},\code{\link[overedge]{as_sf}},\code{\link[overedge]{layer_neatline}}
}
