% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_show_markers.R
\name{make_marker_map}
\alias{make_marker_map}
\alias{make_group_layers}
\alias{prep_markers}
\title{Make marker map}
\usage{
make_marker_map(
  data,
  mapping = ggplot2::aes(),
  prep = TRUE,
  groupname_col = "group",
  groupmeta = NULL,
  style = NULL,
  crs = NULL,
  fn = NULL,
  ...
)

make_group_layers(
  data,
  mapping = ggplot2::aes(),
  groupname_col = "group",
  layers = FALSE,
  ...
)

prep_markers(
  data = NULL,
  groupname_col = "group",
  groupmeta = NULL,
  geocode = FALSE,
  address_col = "address",
  geometry = "point",
  crs = NULL,
  fn = NULL,
  ...
)
}
\arguments{
\item{data}{Data for mapping}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{crs}{coordinate reference system to return}

\item{fn}{Function to apply to data before returning.}

\item{...}{Parameters passed to selected geom}

\item{layers}{defaults to FALSE}
}
\description{
Make marker map

Make group layers

Prep markers for mapping
}
