% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_markers.R
\name{number_markers}
\alias{number_markers}
\title{Sort and number markers}
\usage{
number_markers(
  data,
  groupname_col = NULL,
  number = TRUE,
  sort = "lon",
  desc = FALSE
)
}
\arguments{
\item{data}{Marker data}

\item{groupname_col}{Group column name, Default: NULL}

\item{number}{If TRUE, number the markers, Default: TRUE}

\item{sort_col}{Sort column name, Default: 'lon'}

\item{sort_desc}{default FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Sort and number markers
}
