% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_show_markers.R
\name{prep_markers}
\alias{prep_markers}
\title{Prep markers for mapping}
\usage{
prep_markers(
  data = NULL,
  groupname_col = "group",
  groupmeta = NULL,
  geocode = FALSE,
  address_col = "address",
  geometry = "point",
  crs = NULL,
  fn = NULL,
  ...
)
}
\description{
Prep markers for mapping
}
