% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_marker_key.R
\name{tbl_marker_key}
\alias{tbl_marker_key}
\title{Make a key for a marker map}
\usage{
tbl_marker_key(data, title_col = NULL, groupname_col = NULL, number_col = NULL)
}
\arguments{
\item{data}{Data frame or sf object with marker data}

\item{title_col}{Title column name, Default: NULL}

\item{groupname_col}{Group column name, Default: NULL}

\item{number_col}{Number column name, Default: NULL. If NULL and data does
not contain a column named "number", add a number column created using
\code{dplyr::row_number()} function.}
}
\description{
Designed for use with maps created using \code{layer_show_markers()}
}
