% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_photo_key.R
\name{tbl_photo_key}
\alias{tbl_photo_key}
\title{Create a numbered photo key table}
\usage{
tbl_photo_key(
  data,
  img_width = 320,
  asp = 0.75,
  photo_col = "photo",
  title_col = "title",
  number = FALSE,
  orientation = NULL
)
}
\arguments{
\item{data}{sf object or data frame with photo column containing the file
path or url for photos}

\item{img_width}{Image width.}

\item{photo_col}{Photo path/url column name.}

\item{title_col}{Photo title/description column name.}

\item{img_height}{Image height.}
}
\description{
Create a numbered photo key table
}
