% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_show_markers.R
\name{layer_show_markers}
\alias{layer_show_markers}
\alias{layer_number_markers}
\title{Add a marker layer to a map with or without numbered markers}
\usage{
layer_show_markers(
  data,
  mapping = NULL,
  get = TRUE,
  groupname_col = NULL,
  group_meta = NULL,
  style = NULL,
  crs = NULL,
  fn = NULL,
  number = FALSE,
  number_col = NULL,
  ...
)

layer_number_markers(
  data,
  mapping = NULL,
  number_col = NULL,
  groupname_col = NULL,
  size = 5,
  style = "roundrect",
  geom = "label",
  sort = "lon",
  desc = FALSE,
  ...
)
}
\arguments{
\item{data}{Data with markers, passed to data parameter of \code{\link[overedge:get_location_data]{overedge::get_location_data()}}}

\item{get}{If \code{TRUE}, pass data to get_markers.}

\item{groupname_col}{Group column name, used to join group metadata if group_meta is a non-spatial data frame; Default: \code{NULL}}

\item{group_meta}{Group metadata as a data frame or sf object that intersect with markers; Default: \code{NULL}}

\item{style}{Style; defaults to \code{NULL} for \code{\link[=layer_show_markers]{layer_show_markers()}} (supports
"facet"); defaults to "roundrect" for \code{\link[=layer_number_markers]{layer_number_markers()}},}

\item{crs}{Coordinate reference system for markers, Default: \code{NULL}}

\item{fn}{Function to apply to data before results; gives warning if data is grouped; Default: \code{NULL}}

\item{number}{If \code{TRUE}, number markers using \code{\link[=layer_number_markers]{layer_number_markers()}}}

\item{number_col}{Name of column with numbers; defaults to NULL.}

\item{...}{Additional parameters passed to \code{overedge::layer_location_data}
(include label.size, label.padding, and label.r to define alternate style)}

\item{size}{Marker size, Default: 5}
}
\value{
ggplot2 layers
}
\description{
Add a marker layer to a map with or without numbered markers
}
\examples{
\dontrun{
if (interactive()) {
  ggplot2::ggplot() +
    layer_number_markers(
      data = mapbaltimore::parks,
      groupname_col = "park_district"
    )
}
}
}
