% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_detail_maps.R
\name{make_detail_map}
\alias{make_detail_map}
\alias{make_section_map}
\title{Make detail maps}
\usage{
make_detail_map(
  detail,
  context = NULL,
  fn = NULL,
  groupname_col = NULL,
  palette = NULL,
  aesthetics = "color",
  number = FALSE,
  overlay = NULL,
  paper = NULL,
  orientation = NULL,
  context_dist = 0,
  detail_diag_ratio = 0.1,
  access_token = Sys.getenv("MAPBOX_SECRET_TOKEN"),
  columns = 1,
  sort = "lon",
  rows = 1,
  scale = 1,
  neatline = TRUE,
  neatline_dist = 0,
  expand = TRUE,
  crs = NULL,
  inset_legend = TRUE,
  legend_position = NULL,
  ...
)

make_section_map(
  section,
  groupname_col = NULL,
  overlay = NULL,
  fn = NULL,
  sort = NULL,
  access_token = NULL,
  scale = 0.6,
  asp = NULL,
  number = FALSE,
  neatline = TRUE,
  neatline_dist = 0,
  expand = TRUE,
  crs = NULL,
  ...
)
}
\arguments{
\item{detail}{an sf object (or object that is convertible with
\code{\link[overedge:as_sf]{overedge::as_sf()}}), list of sf objects, or data frame with a sf list
column named data}

\item{context}{Broader context area containing all detail maps, Default: NULL}

\item{fn}{Function, Default: NULL}

\item{number}{If TRUE, number data, Default: FALSE}

\item{overlay}{Overlay, Default: NULL}

\item{access_token}{Mapbox access token, Default: NULL}

\item{scale}{Scale, Default: 0.6}

\item{...}{Additional parameters used by various functions.}

\item{section}{Section (sf obect)}

\item{asp}{Aspect ratio, Default: NULL}
}
\description{
Make detail maps, e.g. nodes of development within a broader area, examples
of a type using data created by using \code{\link[dplyr:group_by]{dplyr::group_by()}} and
\code{\link[dplyr:group_nest]{dplyr::group_nest()}} if the data is an sf object and keep = TRUE.
}
\details{
Make section map:

All parameters that are specific to individual section maps should be passed
to sect_params. asp is automatically calculated based on the width and height
of the overall paper, and the number of rows and columns specified in
make_detail_map.

Some make_detail_map parameters are passed on directly to make_section_map
including:
\itemize{
\item fn (applied to each section before mapping - may be useful for recoding labels)
\item access_token and scale - used to create section base map
\item number - indicator of whether to call layer_number_markers and add to layers for section map
}

All other parameters are defined by the dots ... Supported options for ...
are passed to one of the following functions:
\itemize{
\item layer_number_markers: sort, number_col, and group_col parameters
\item \link[overedge:theme_ext]{overedge::theme_legend}: all parameters (except method) area supported but these
parameters must be passed as a list so they can be called in the following
way, e.g. sect_params$legend$position.
\item \link[overedge:layer_neatline]{overedge::layer_neatline}: the expand parameter for  and the parameter neatline_dist to
allow the adjustment of the section neatline (needed for the expand parameter
to work as expected)
}

All of these variables are currently defined globally and cannot be modified
for detail maps with the section (unless you call make_section_map directly).

DETAILS
}
