% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_markers.R
\name{number_markers}
\alias{number_markers}
\title{Sort and number markers}
\usage{
number_markers(
  data,
  groupname_col = NULL,
  number = TRUE,
  sort = "lon",
  desc = NULL
)
}
\arguments{
\item{data}{Marker data}

\item{groupname_col}{Group column name, Default: \code{NULL}}

\item{number}{If TRUE, number the markers, Default: \code{TRUE}}

\item{desc}{If \code{TRUE}, sort descending; set to \code{FALSE} if sort is "lon" and
TRUE is sort is "lat" or any other value, default \code{NULL}}

\item{sort_col}{Sort column name, Default: 'lon'}
}
\value{
A sf object with a number column ordered by sort.
}
\description{
Used with \link{layer_number_markers}
}
