% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_marker_key.R
\name{tbl_marker_key}
\alias{tbl_marker_key}
\title{Make a key for a marker map}
\usage{
tbl_marker_key(
  data,
  title_col = NULL,
  groupname_col = NULL,
  number_col = NULL,
  color = FALSE,
  palette = NULL
)
}
\arguments{
\item{data}{Data frame or sf object with marker data}

\item{title_col}{Title column name, Default: \code{NULL}}

\item{groupname_col}{Group column name, Default: \code{NULL}}

\item{number_col}{Number column name, Default: \code{NULL}. If \code{NULL} and data does
not contain a column named "number", add a number column created using
\code{\link[dplyr:ranking]{dplyr::row_number()}} function.}

\item{color}{If \code{TRUE}, apply a cell fill color to the group headings in the
table (defined by groupname_col)}

\item{palette}{palette to use for the group heading fill colors (passed to
\code{\link[=group_data_pal]{group_data_pal()}})}
}
\description{
Designed for use with maps created using \code{layer_show_markers()}
}
